Merge branch 'feature/kunen1/nested2' of https://github.com/LLNL/RAJA into feature/kunen1/nested2

Conflicts:
	include/RAJA/policy/cuda/nested/CudaKernel.hpp
	include/RAJA/policy/cuda/nested/For.hpp

+++ include/RAJA/policy/cuda/nested/For.hpp
@@ -8,7 +8,6 @@
  ******************************************************************************
  */
  
-
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 // Copyright (c) 2016-18, Lawrence Livermore National Security, LLC.
 // 
@@ -36,86 +35,96 @@ namespace RAJA
 {
 
 
-
 namespace nested
 {
 
 
-namespace internal{
-
+namespace internal
+{
 
 
 /*
  * Executor for thread work sharing loop inside a Cuda Kernel.
  *
  */
-template <typename Data, camp::idx_t ArgumentId, typename... EnclosedStmts, typename IndexCalc>
-struct CudaStatementExecutor<Data, For<ArgumentId, cuda_thread_exec, EnclosedStmts...>, IndexCalc> {
+template <typename Data,
+          camp::idx_t ArgumentId,
+          typename... EnclosedStmts,
+          typename IndexCalc>
+struct CudaStatementExecutor<Data,
+                             For<ArgumentId,
+                                 cuda_thread_exec,
+                                 EnclosedStmts...>,
+                             IndexCalc> {
 
   using stmt_list_t = StatementList<EnclosedStmts...>;
-  using index_calc_t = ExtendCudaIndexCalc<IndexCalc,CudaIndexCalc_Policy<ArgumentId, cuda_thread_exec>>;
+  using index_calc_t =
+      ExtendCudaIndexCalc<IndexCalc,
+                          CudaIndexCalc_Policy<ArgumentId, cuda_thread_exec>>;
 
-  using enclosed_stmts_t = CudaStatementListExecutor<Data, stmt_list_t, index_calc_t>;
+  using enclosed_stmts_t =
+      CudaStatementListExecutor<Data, stmt_list_t, index_calc_t>;
   enclosed_stmts_t enclosed_stmts;
 
-  inline
-  RAJA_DEVICE
-  void exec(Data &data, int num_logical_blocks, int block_carry)
+  inline RAJA_DEVICE void exec(Data &data,
+                               int num_logical_blocks,
+                               int block_carry)
   {
     // execute enclosed statements
     enclosed_stmts.exec(data, num_logical_blocks, block_carry);
   }
 
 
-  inline
-  RAJA_DEVICE
-  void initBlocks(Data &data, int num_logical_blocks, int block_stride)
+  inline RAJA_DEVICE void initBlocks(Data &data,
+                                     int num_logical_blocks,
+                                     int block_stride)
   {
     enclosed_stmts.initBlocks(data, num_logical_blocks, block_stride);
   }
 
   RAJA_INLINE
-	LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical){
-		
-		LaunchDim dim = enclosed_stmts.calculateDimensions(data, max_physical);
-		
-		dim.threads *= segment_length<ArgumentId>(data);
-		
-		return dim;
-	}
+  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical)
+  {
+
+    LaunchDim dim = enclosed_stmts.calculateDimensions(data, max_physical);
+
+    dim.threads *= segment_length<ArgumentId>(data);
+
+    return dim;
+  }
 };
-
-
-
-
-
-
 
 
 /*
  * Executor for block work sharing loop inside a Cuda Kernel.
  *
  */
-template <typename Data, camp::idx_t ArgumentId, typename... EnclosedStmts, typename IndexCalc>
-struct CudaStatementExecutor<Data, For<ArgumentId, cuda_block_exec, EnclosedStmts...>, IndexCalc> : public CudaBlockLoop<ArgumentId, 1>{
+template <typename Data,
+          camp::idx_t ArgumentId,
+          typename... EnclosedStmts,
+          typename IndexCalc>
+struct CudaStatementExecutor<Data,
+                             For<ArgumentId, cuda_block_exec, EnclosedStmts...>,
+                             IndexCalc> : public CudaBlockLoop<ArgumentId, 1> {
 
   using stmt_list_t = StatementList<EnclosedStmts...>;
 
-  using enclosed_stmts_t = CudaStatementListExecutor<Data, stmt_list_t, IndexCalc>;
+  using enclosed_stmts_t =
+      CudaStatementListExecutor<Data, stmt_list_t, IndexCalc>;
   enclosed_stmts_t enclosed_stmts;
 
 
-  inline
-  RAJA_DEVICE
-  void exec(Data &data, int num_logical_blocks, int block_carry)
+  inline RAJA_DEVICE void exec(Data &data,
+                               int num_logical_blocks,
+                               int block_carry)
   {
     execBlockLoop(*this, data, num_logical_blocks, block_carry);
   }
 
 
-  inline
-  RAJA_DEVICE
-  void initBlocks(Data &data, int num_logical_blocks, int block_stride)
+  inline RAJA_DEVICE void initBlocks(Data &data,
+                                     int num_logical_blocks,
+                                     int block_stride)
   {
     int len = segment_length<ArgumentId>(data);
     initBlockLoop(enclosed_stmts, data, len, num_logical_blocks, block_stride);
@@ -123,76 +132,79 @@ struct CudaStatementExecutor<Data, For<ArgumentId, cuda_block_exec, EnclosedStmt
 
 
   RAJA_INLINE
-	LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical){
+  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical)
+  {
 
     LaunchDim dim = enclosed_stmts.calculateDimensions(data, max_physical);
 
-		dim.blocks *= segment_length<ArgumentId>(data);
+    dim.blocks *= segment_length<ArgumentId>(data);
 
-		return dim;
-	}
+    return dim;
+  }
 };
-
-
-
-
-
-
 
 
 /*
  * Executor for thread and block work sharing loop inside a Cuda Kernel.
  *
  */
-template <typename Data, camp::idx_t ArgumentId, size_t max_threads, typename... EnclosedStmts, typename IndexCalc>
-struct CudaStatementExecutor<Data, For<ArgumentId, cuda_threadblock_exec<max_threads>, EnclosedStmts...>, IndexCalc> : public CudaBlockLoop<ArgumentId, max_threads> {
+template <typename Data,
+          camp::idx_t ArgumentId,
+          size_t max_threads,
+          typename... EnclosedStmts,
+          typename IndexCalc>
+struct CudaStatementExecutor<Data,
+                             For<ArgumentId,
+                                 cuda_threadblock_exec<max_threads>,
+                                 EnclosedStmts...>,
+                             IndexCalc>
+    : public CudaBlockLoop<ArgumentId, max_threads> {
 
   using stmt_list_t = StatementList<EnclosedStmts...>;
-  using index_calc_t = ExtendCudaIndexCalc<IndexCalc,CudaIndexCalc_Policy<ArgumentId, cuda_thread_exec>>;
+  using index_calc_t =
+      ExtendCudaIndexCalc<IndexCalc,
+                          CudaIndexCalc_Policy<ArgumentId, cuda_thread_exec>>;
 
-  using enclosed_stmts_t = CudaStatementListExecutor<Data, stmt_list_t, index_calc_t>;
+  using enclosed_stmts_t =
+      CudaStatementListExecutor<Data, stmt_list_t, index_calc_t>;
   enclosed_stmts_t enclosed_stmts;
 
 
-  inline
-  RAJA_DEVICE
-  void exec(Data &data, int num_logical_blocks, int block_carry)
+  inline RAJA_DEVICE void exec(Data &data,
+                               int num_logical_blocks,
+                               int block_carry)
   {
     execBlockLoop(*this, data, num_logical_blocks, block_carry);
   }
 
 
-  inline
-  RAJA_DEVICE
-  void initBlocks(Data &data, int num_logical_blocks, int block_stride)
+  inline RAJA_DEVICE void initBlocks(Data &data,
+                                     int num_logical_blocks,
+                                     int block_stride)
   {
     int len = segment_length<ArgumentId>(data);
     initBlockLoop(enclosed_stmts, data, len, num_logical_blocks, block_stride);
   }
 
-	RAJA_INLINE
-  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical){
+  RAJA_INLINE
+  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical)
+  {
 
     LaunchDim dim = enclosed_stmts.calculateDimensions(data, max_physical);
 
-		// Compute how many blocks
-		int len = segment_length<ArgumentId>(data);
-		int num_blocks = len / max_threads;
-		if(num_blocks*max_threads < len){
-			num_blocks ++;
-		}
+    // Compute how many blocks
+    int len = segment_length<ArgumentId>(data);
+    int num_blocks = len / max_threads;
+    if (num_blocks * max_threads < len) {
+      num_blocks++;
+    }
 
-		dim.blocks *= num_blocks;
-		dim.threads *= (int)max_threads;
+    dim.blocks *= num_blocks;
+    dim.threads *= max_threads;
 
-		return dim;
-	}
+    return dim;
+  }
 };
-
-
-
-
-
 
 
 /*
@@ -200,55 +212,71 @@ struct CudaStatementExecutor<Data, For<ArgumentId, cuda_threadblock_exec<max_thr
  *
  * This is specialized since it need to execute the loop immediately.
  */
-template <typename Data, camp::idx_t ArgumentId, typename... EnclosedStmts, typename IndexCalc>
-struct CudaStatementExecutor<Data, For<ArgumentId, seq_exec, EnclosedStmts...>, IndexCalc> {
+template <typename Data,
+          camp::idx_t ArgumentId,
+          typename... EnclosedStmts,
+          typename IndexCalc>
+struct CudaStatementExecutor<Data,
+                             For<ArgumentId, seq_exec, EnclosedStmts...>,
+                             IndexCalc> {
 
   using stmt_list_t = StatementList<EnclosedStmts...>;
-  using index_calc_t = ExtendCudaIndexCalc<IndexCalc,CudaIndexCalc_Policy<ArgumentId, seq_exec>>;
+  using index_calc_t =
+      ExtendCudaIndexCalc<IndexCalc,
+                          CudaIndexCalc_Policy<ArgumentId, seq_exec>>;
 
-  using enclosed_stmts_t = CudaStatementListExecutor<Data, stmt_list_t, index_calc_t>;
+  using enclosed_stmts_t =
+      CudaStatementListExecutor<Data, stmt_list_t, index_calc_t>;
   enclosed_stmts_t enclosed_stmts;
 
-  inline
-  RAJA_DEVICE
-  void exec(Data &data, int num_logical_blocks, int block_carry)
+  inline RAJA_DEVICE void exec(Data &data,
+                               int num_logical_blocks,
+                               int block_carry)
   {
     // execute enclosed statements
     enclosed_stmts.exec(data, num_logical_blocks, block_carry);
   }
 
 
-  inline
-  RAJA_DEVICE
-  void initBlocks(Data &data, int num_logical_blocks, int block_stride)
+  inline RAJA_DEVICE void initBlocks(Data &data,
+                                     int num_logical_blocks,
+                                     int block_stride)
   {
     enclosed_stmts.initBlocks(data, num_logical_blocks, block_stride);
   }
 
   RAJA_INLINE
-  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical){
+  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical)
+  {
 
     return enclosed_stmts.calculateDimensions(data, max_physical);
-
   }
 };
 
 
-template <typename Data, camp::idx_t ArgumentId, typename... EnclosedStmts, typename Segments>
-struct CudaStatementExecutor<Data, For<ArgumentId, seq_exec, EnclosedStmts...>, CudaIndexCalc_Terminator<Segments>> {
+template <typename Data,
+          camp::idx_t ArgumentId,
+          typename... EnclosedStmts,
+          typename Segments>
+struct CudaStatementExecutor<Data,
+                             For<ArgumentId, seq_exec, EnclosedStmts...>,
+                             CudaIndexCalc_Terminator<Segments>> {
 
   using stmt_list_t = StatementList<EnclosedStmts...>;
 
-  using enclosed_stmts_t = CudaStatementListExecutor<Data, stmt_list_t, CudaIndexCalc_Terminator<Segments>>;
+  using enclosed_stmts_t =
+      CudaStatementListExecutor<Data,
+                                stmt_list_t,
+                                CudaIndexCalc_Terminator<Segments>>;
   enclosed_stmts_t enclosed_stmts;
 
-  inline
-  RAJA_DEVICE
-  void exec(Data &data, int num_logical_blocks, int block_carry)
+  inline RAJA_DEVICE void exec(Data &data,
+                               int num_logical_blocks,
+                               int block_carry)
   {
     int len = segment_length<ArgumentId>(data);
 
-    for(int i = 0;i < len;++ i){
+    for (int i = 0; i < len; ++i) {
       data.template assign_offset<ArgumentId>(i);
 
       // execute enclosed statements
@@ -257,22 +285,20 @@ struct CudaStatementExecutor<Data, For<ArgumentId, seq_exec, EnclosedStmts...>, 
   }
 
 
-  inline
-  RAJA_DEVICE
-  void initBlocks(Data &data, int num_logical_blocks, int block_stride)
+  inline RAJA_DEVICE void initBlocks(Data &data,
+                                     int num_logical_blocks,
+                                     int block_stride)
   {
     enclosed_stmts.initBlocks(data, num_logical_blocks, block_stride);
   }
 
   RAJA_INLINE
-  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical){
+  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical)
+  {
 
     return enclosed_stmts.calculateDimensions(data, max_physical);
-
   }
 };
-
-
 
 
 /*
@@ -280,21 +306,29 @@ struct CudaStatementExecutor<Data, For<ArgumentId, seq_exec, EnclosedStmts...>, 
  *
  * This is specialized since it need to execute the loop immediately.
  */
-template <typename Data, camp::idx_t ArgumentId, typename... EnclosedStmts, typename IndexCalc>
-struct CudaStatementExecutor<Data, For<ArgumentId, cuda_seq_syncthreads_exec, EnclosedStmts...>, IndexCalc> {
+template <typename Data,
+          camp::idx_t ArgumentId,
+          typename... EnclosedStmts,
+          typename IndexCalc>
+struct CudaStatementExecutor<Data,
+                             For<ArgumentId,
+                                 cuda_seq_syncthreads_exec,
+                                 EnclosedStmts...>,
+                             IndexCalc> {
 
   using stmt_list_t = StatementList<EnclosedStmts...>;
 
-  using enclosed_stmts_t = CudaStatementListExecutor<Data, stmt_list_t, IndexCalc>;
+  using enclosed_stmts_t =
+      CudaStatementListExecutor<Data, stmt_list_t, IndexCalc>;
   enclosed_stmts_t enclosed_stmts;
 
-  inline
-  RAJA_DEVICE
-  void exec(Data &data, int num_logical_blocks, int block_carry)
+  inline RAJA_DEVICE void exec(Data &data,
+                               int num_logical_blocks,
+                               int block_carry)
   {
     int len = segment_length<ArgumentId>(data);
 
-    for(int i = 0;i < len;++ i){
+    for (int i = 0; i < len; ++i) {
       data.template assign_offset<ArgumentId>(i);
 
       // execute enclosed statements
@@ -305,29 +339,25 @@ struct CudaStatementExecutor<Data, For<ArgumentId, cuda_seq_syncthreads_exec, En
   }
 
 
-  inline
-  RAJA_DEVICE
-  void initBlocks(Data &data, int num_logical_blocks, int block_stride)
+  inline RAJA_DEVICE void initBlocks(Data &data,
+                                     int num_logical_blocks,
+                                     int block_stride)
   {
     enclosed_stmts.initBlocks(data, num_logical_blocks, block_stride);
   }
 
   RAJA_INLINE
-  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical){
+  LaunchDim calculateDimensions(Data const &data, LaunchDim const &max_physical)
+  {
 
     return enclosed_stmts.calculateDimensions(data, max_physical);
-
   }
 };
 
 
-
-
-
-} // namespace internal
+}  // namespace internal
 }  // end namespace nested
 }  // end namespace RAJA
-
 
 
 #endif /* RAJA_pattern_nested_HPP */
